MM4 compatibility
