+++ Gemfile
@@ -3,18 +3,27 @@ source "http://rubygems.org"
 # Specify your gem's dependencies in middleman-thumbnailer.gemspec
 gemspec
 
-gem "middleman", git: 'https://github.com/middleman/middleman.git', branch: 'v3-stable'
-
-
-group :development, :test do
-  # Test tools
-  gem "cucumber", "~> 1.3.1"
-  gem "fivemat"
-  gem "aruba",    "~> 0.5.1"
-  gem "rspec",    "~> 2.12"
-  gem "builder",  "~> 3.0.0"
-  gem "erubis"
-
-  # gem "pry"
-  # gem "pry-plus", git: 'git@github.com:nhemsley/pry-plus.git'
+group :development do
+  gem 'rake'
+  gem 'rdoc'
+  gem 'yard'
 end
+
+group :test do
+  gem 'cucumber'
+  gem 'aruba'
+  gem 'rspec'
+  gem "fivemat"
+end
+
+# group :development, :test do
+#   # Test tools
+#   gem "cucumber", "~> 1.3.1"
+#   gem "aruba",    "~> 0.5.1"
+#   gem "rspec",    "~> 2.12"
+#   gem "builder",  "~> 3.0.0"
+#   gem "erubis"
+#
+#   # gem "pry"
+#   # gem "pry-plus", git: 'git@github.com:nhemsley/pry-plus.git'
+# end
