+++ lib/middleman-thumbnailer.rb
@@ -1,8 +1,6 @@
 require "middleman-core"
 
-require "middleman-thumbnailer/version"
-
-::Middleman::Extensions.register(:thumbnailer) do
+Middleman::Extensions.register :thumbnailer do
   require "middleman-thumbnailer/extension"
-  ::Middleman::Thumbnailer
+  ::Middleman::Thumbnailer::Extension
 end
