--- lib/middleman_extension.rb
@@ -1,1 +1,0 @@
-require "middleman-thumbnailer"
